Clean Cordova map and array implementations
