+++ other-platforms-support/cordova/android/src/main/java/com/wultra/android/powerauth/cdv/util/Arguments.java
@@ -26,10 +26,6 @@ public class Arguments {
       return makeNativeMap((Map<String, Object>) object);
     } else if (object instanceof Bundle) {
       return makeNativeMap((Bundle) object);
-    // } else if (object instanceof JavaOnlyMap) {
-    //   return makeNativeMap(((JavaOnlyMap) object).toHashMap());
-    // } else if (object instanceof JavaOnlyArray) {
-    //   return makeNativeArray(((JavaOnlyArray) object).toArrayList());
     } else {
       // Boolean, Integer, Double, String, WritableNativeArray, WritableNativeMap
       return object;
